Optimizes ngram analysis performance
Improves performance of ngram analysis by switching from u64 to f32 for ngram counts.
Uses the fast-math crate for faster log calculations in the pl function.
This significantly speeds up the Metropolis-Hastings algorithm.

+++ src/analysis/fast_ngram_generator.rs
@@ -1,11 +1,11 @@
-pub fn fast_ngram_generator(input: &str) -> [[u64; 26]; 26] {
-    let mut ngram: [[u64; 26]; 26] = [[0; 26]; 26];
+pub fn fast_ngram_generator(input: &str) -> [[f32; 26]; 26] {
+    let mut ngram: [[f32; 26]; 26] = [[0.0; 26]; 26];
 
     input
         .as_bytes()
         .windows(2)
         .for_each(|w| {
-            ngram[(w[0]-b'A') as usize][(w[1]-b'A') as usize] += 1;
+            ngram[(w[0]-b'A') as usize][(w[1]-b'A') as usize] += 1.0;
         });
 
     ngram
